Improved how overlay is painted
I've moved the code to scale the overlay graphics contexts into its own method, along with creating a method to configure the graphics contexts its using. This second method makes the rendering and alpha interpolation prioritize quality over speed, and is applied to both the image that the overlay is rendered to and a copy of the graphics context given to the method. I've also made it so that the antialiasing is set for text masks based off the mask painter's antialiasing.