Reworked the spiral interpolating code
I realized that, since we're using quadratic bezier curves this time, we don't need to calculate the points on the segment of the arc at 1/3 and 2/3rds of the length of the arc. Instead, I can just calculate one point on the segment of the arc at any given point and use a method from GeometryMath to calculate the control point of the quadratic bezier curve that passes through the three points. This greatly simplifies any logic that I would have needed to implement.