Added settings interface for overlay mask image
I've added an interface to get the settings for image overlay masks. I've also made it so that the settings for the image overlay mask are now handled by a dedicated interface, of which an instance is returned by the settings interface for the whole program. I've added implementations of said interface to SpiralGeneratorConfig and SpiralGeneratorProperties. Additionally I've deprecated the old settings methods and constants that relate to the image overlay mask settings and made it so those methods now forward the calls to the dedicated interface object.