Added settings interface for overlay mask messages
I've added an interface to get the settings for multiple messages overlay masks. I've also made it so that the settings for the multiple messages overlay mask are now handled by a dedicated interface, of which an instance is returned by the settings interface for the whole program. I've added implementations of said interface to SpiralGeneratorConfig and SpiralGeneratorProperties. Additionally I've deprecated the old settings methods and constants that relate to the multiple messages overlay mask settings and made it so those methods now forward the calls to the dedicated interface object.